AbstractMessage, Http\Request, Http\Response
