+++ library/Zend/Http/Response.php
@@ -10,12 +10,17 @@
 
 namespace Zend\Http;
 
-use Zend\Stdlib\Message;
 use Zend\Stdlib\ResponseInterface;
 
-class Response extends Message implements ResponseInterface
+/**
+ * HTTP Response
+ *
+ * @category  Zend
+ * @package   Zend_Http
+ * @link      http://www.w3.org/Protocols/rfc2616/rfc2616-sec6.html#sec6
+ */
+class Response extends AbstractMessage implements ResponseInterface
 {
-
     /**#@+
      * @const int Status codes
      */
@@ -77,20 +82,7 @@ class Response extends Message implements ResponseInterface
     const STATUS_CODE_507 = 507;
     const STATUS_CODE_508 = 508;
     const STATUS_CODE_511 = 511;
-
     /**#@-*/
-
-    /**#@+
-     * @const string Version constant numbers
-     */
-    const VERSION_11 = '1.1';
-    const VERSION_10 = '1.0';
-    /**#@-*/
-
-    /**
-     * @var string
-     */
-    protected $version = self::VERSION_11;
 
     /**
      * @var array Recommended Reason Phrases
@@ -171,15 +163,11 @@ class Response extends Message implements ResponseInterface
     protected $reasonPhrase = null;
 
     /**
-     * @var Headers
-     */
-    protected $headers = null;
-
-    /**
      * Populate object from string
      *
      * @param  string $string
      * @return Response
+     * @throws Exception\InvalidArgumentException
      */
     public static function fromString($string)
     {
@@ -191,9 +179,13 @@ class Response extends Message implements ResponseInterface
         $firstLine = array_shift($lines);
 
         $response = new static();
-        $matches = null;
-        if (!preg_match('/^HTTP\/(?P<version>1\.[01]) (?P<status>\d{3})(?:[ ]+(?P<reason>.+))?$/', $firstLine, $matches)) {
-            throw new Exception\InvalidArgumentException('A valid response status line was not found in the provided string');
+
+        $regex   = '/^HTTP\/(?P<version>1\.[01]) (?P<status>\d{3})(?:[ ]+(?P<reason>.+))?$/';
+        $matches = array();
+        if (!preg_match($regex, $firstLine, $matches)) {
+            throw new Exception\InvalidArgumentException(
+                'A valid response status line was not found in the provided string'
+            );
         }
 
         $response->version = $matches['version'];
@@ -215,9 +207,9 @@ class Response extends Message implements ResponseInterface
                 continue;
             }
             if ($isHeader) {
-                $headers[] .= $nextLine;
+                $headers[] = $nextLine;
             } else {
-                $content[] .= $nextLine;
+                $content[] = $nextLine;
             }
         }
 
@@ -233,47 +225,6 @@ class Response extends Message implements ResponseInterface
     }
 
     /**
-     * Render the status line header
-     *
-     * @return string
-     */
-    public function renderStatusLine()
-    {
-        $status = sprintf(
-            'HTTP/%s %d %s',
-            $this->getVersion(),
-            $this->getStatusCode(),
-            $this->getReasonPhrase()
-        );
-        return trim($status);
-    }
-
-    /**
-     * Set response headers
-     *
-     * @param  Headers $headers
-     * @return Response
-     */
-    public function setHeaders(Headers $headers)
-    {
-        $this->headers = $headers;
-        return $this;
-    }
-
-    /**
-     * Get response headers
-     *
-     * @return Headers
-     */
-    public function getHeaders()
-    {
-        if ($this->headers === null || is_string($this->headers)) {
-            $this->headers = (is_string($this->headers)) ? Headers::fromString($this->headers) : new Headers();
-        }
-        return $this->headers;
-    }
-
-    /**
      * @return Header\SetCookie[]
      */
     public function getCookie()
@@ -282,21 +233,24 @@ class Response extends Message implements ResponseInterface
     }
 
     /**
-     * @param string $version
+     * Set HTTP status code and (optionally) message
+     *
+     * @param  integer $code
+     * @throws Exception\InvalidArgumentException
      * @return Response
      */
-    public function setVersion($version)
+    public function setStatusCode($code)
     {
-        $this->version = $version;
+        $const = get_called_class() . '::STATUS_CODE_' . $code;
+        if (!is_numeric($code) || !defined($const)) {
+            $code = is_scalar($code) ? $code : gettype($code);
+            throw new Exception\InvalidArgumentException(sprintf(
+                'Invalid status code provided: "%s"',
+                $code
+            ));
+        }
+        $this->statusCode = (int) $code;
         return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getVersion()
-    {
-        return $this->version;
     }
 
     /**
@@ -330,27 +284,6 @@ class Response extends Message implements ResponseInterface
             return $this->recommendedReasonPhrases[$this->statusCode];
         }
         return $this->reasonPhrase;
-    }
-
-    /**
-     * Set HTTP status code and (optionally) message
-     *
-     * @param integer $code
-     * @throws Exception\InvalidArgumentException
-     * @return \Zend\Http\Response
-     */
-    public function setStatusCode($code)
-    {
-        $const = get_called_class() . '::STATUS_CODE_' . $code;
-        if (!is_numeric($code) || !defined($const)) {
-            $code = is_scalar($code) ? $code : gettype($code);
-            throw new Exception\InvalidArgumentException(sprintf(
-                'Invalid status code provided: "%s"',
-                $code
-            ));
-        }
-        $this->statusCode = (int) $code;
-        return $this;
     }
 
     /**
@@ -469,6 +402,21 @@ class Response extends Message implements ResponseInterface
         return (200 <= $code && 300 > $code);
     }
 
+    /**
+     * Render the status line header
+     *
+     * @return string
+     */
+    public function renderStatusLine()
+    {
+        $status = sprintf(
+            'HTTP/%s %d %s',
+            $this->getVersion(),
+            $this->getStatusCode(),
+            $this->getReasonPhrase()
+        );
+        return trim($status);
+    }
 
     /**
      * Render entire response as HTTP response string
@@ -485,20 +433,11 @@ class Response extends Message implements ResponseInterface
     }
 
     /**
-     * Implements magic __toString()
-     *
-     * @return string
-     */
-    public function __toString()
-    {
-        return $this->toString();
-    }
-
-    /**
      * Decode a "chunked" transfer-encoded body and return the decoded text
      *
-     * @param string $body
+     * @param  string $body
      * @return string
+     * @throws Exception\RuntimeException
      */
     protected function decodeChunkedBody($body)
     {
@@ -525,8 +464,9 @@ class Response extends Message implements ResponseInterface
      *
      * Currently requires PHP with zlib support
      *
-     * @param string $body
+     * @param  string $body
      * @return string
+     * @throws Exception\RuntimeException
      */
     protected function decodeGzip($body)
     {
@@ -544,8 +484,9 @@ class Response extends Message implements ResponseInterface
      *
      * Currently requires PHP with zlib support
      *
-     * @param string $body
+     * @param  string $body
      * @return string
+     * @throws Exception\RuntimeException
      */
     protected function decodeDeflate($body)
     {
